Fix : SkillCompo

+++ Assets/Member/Kmj/01.Scipt/Player/MousePlayerEnergy.cs
@@ -16,7 +16,7 @@ public class MousePlayerEnergy : MonoBehaviour
 
     public void UseEnergyTimeAtTime(int useEnergy)
     {
-        energy -= useEnergy * (int)Time.deltaTime;
+        energy -= useEnergy * Time.deltaTime;
         print((int)Time.deltaTime);
     }
